requirements.txt
pypdf>=4.0.0
openai>=1.0.0
anthropic>=0.20.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
